requirement.txt
torch
torchvision
numpy
opencv-python
Pillow
tensorboard
yacs
tqdm
wandb
scikit-image
pandas
Ninja
timm
matplotlib